Remove message filter since it causes problems
http://www.autohotkey.com/forum/viewtopic.php?p=358029#358029

Blocking input and/or paint messages at that point (even if it improves
the reliability of subclassing) should not be done as it makes modal
dialogs such as MsgBox and FileSelectFile unresponsive.  If blocking
messages there helps subclassing despite the use of Critical n, perhaps
messages are still getting through somehow and that should be fixed.