Add a new #InputGroup directive for controlling which generated inputs get ignored.
To avoid infinite loops and other side effects, AHK ignores self-generated input events.
It does this by setting a sentinel value in the ExtraInfo field when sending input. In
previous versions, a single sentinel value was used for all scripts, making it impossible
for a script to send input that triggers a hotkey in another script.

The #InputGroup directive provides a way to change the sentinel, so that each script only
ignores input from the other scripts in the same input group.