Debugger: Fixed property_get/value data truncation and error-handling.
If the data exceeds max_data, truncate the data but return the total
size in the size attribute. This and some other cases previously caused
the response to indicate the property is empty.

If a memory allocation or string conversion fails, return an error 998
response instead of an incomplete (invalid) success response.