Re-Added the new ImageSearch method.
Added a new ImageSearch method that by default supports transparency
(alpha of 0) in
needle images (PNG, JPG, TIF, GIF, ICO) and in some cases can be much
faster than the old ImageSearch method.

Added the following internal functions: ImageSearch2_EraseTransColor(),
ImageSearch2_PixelAverageScan(), ImageSearch2_WithVariance(),
ImageSearch2_NoVariance().

Modified the ImageSearch function to use the new search functions.

Added the following options to the ImageSearch command:

*M2: Uses a different (quite possibly faster) method to search for the
image. This method requires GDIPlus which is included in XP and later
operating systems. GDIPlus can also be added to older operating systems
by copying the DLL. If GDIPlus is not available searches will be
unreliable and may fail completely or match anything. If any other
number is used the default (normal) search method will be used. For a
full list of differences see the Mode 2 section below.
*SDn: The scan direction to use when matching the needle image. Specify
for n a number between 0 and 4 (inclusive) to indicate the scan
direction that should be used. Depending on the needle image, one scan
direction over another can be the difference between a 40 second search
time and a .2 second search time. The available options for scan
direction are: 0 (default): Auto detect the “best” scan direction. 1:
Scan left->right, top->bottom. 2: Scan left->right, bottom->top. 3: Scan
right->left, bottom->top. 4: Scan right->left, top->bottom. If the value
is not set or is set to 0 and a “best” scan direction can’t be found it
will failover to 1. The scan direction used does not affect where on the
screen the needle image is matched because the image is always matched
relative to the current search position. If scan direction is not set or
is set to 0 and the needle image is smaller than 6x6 it will default to
the scan direction of 1. This option is only used if the M2 search
method is used.

Mode 2 differences

Mode 2 brings possibly breaking changes over the default (normal) search
method.
1. It will use the pure-transparent alpha values of an image if they
exist without the need to specify a *Trans value.
2. While using the mode 2 search method the *Trans value cannot contain
an alpha value ([a]RGB).
3. With variation (depending on the needle image and level of variation)
searches are much faster.
4. The scan direction used to match the needle image to the haystack
(screen) can be manually set or automatically chosen (see *SDn above).
5. If the scan direction is not manually set and the needle image is
larger than 6x6 a unique pixel is chosen from the needle image as a
match-first location when looking for the needle image on the haystack
(screen). With variation and depending on the needle image, this can
dramatically reduce the time it takes to match or fail to match the
needle image.