Revised handling of named actions for consistency and to fix bugs.
User-defined commands are now handled the same as built-in commands,
in that ternary and post-inc/dec no longer take priority (cmd ?x:y).
msgobx,() was treated as an expression instead of an error.
msgobx,[] as above.
n,++ was treated as an expression (which silently failed).
fun () was treated as an expression instead of a user-defined command.