Fixed "%i-1%" to not interpret 1 as a variable name.
The check for g_DerefChar was intended to detect 0%i% as an (invalid)
double-deref, but in that case the error message only appeared when
the double-deref was evaluated.  Without the check, the error message
indicates a missing space.  This might be confusing if auto-concat was
not intended, but it doesn't seem worth the extra code to detect it as
an invalid double-deref at load-time.