Tweaked COM interface behaviour.
X.Y in VBScript and C# now returns X.Y if it is NOT a Func object;
previously, it attempted to call X.Y() regardless of type.

X(Y) in VBScript, JScript and C# now returns X[Y] unless X is a Func
object, in which case it is called as before.

X[Y] in C# now returns X[Y] if X is NOT a Func object;
previously, it attempted to call X[Y]() regardless of type.

X.Y() in C# now returns X.Y if it is NOT a Func object,
due to ambiguity in the C# dispatch implementation.