fixed from PVS-Studio:
V728 An excessive check can be simplified. The '(A && B) || (!A && !B)' expression is equivalent to the 'bool(A) == bool(B)' expression. wingroup.cpp 114
V773 The function was exited without releasing the 'new_mem' pointer. A memory leak is possible. keyboard_mouse.cpp 2712
V773 The function was exited without releasing the 'token' pointer. A memory leak is possible. script.cpp 12300
V728 An excessive check can be simplified. The '(A && B) || (!A && !B)' expression is equivalent to the 'bool(A) == bool(B)' expression. script.h 1987