ComObjActive() looks in the ROT for instances identified by Monikers
Now it is possible to get objects from the Running Object Table (ROT)
by Moniker even for names which are not present in the Registry.

CLSIDFromString() fails if the human readable name is not present
in the Registry. Now if CLSIDFromString() fails, the ROT is queried
for objects registered with the requested name.