StdIn Scripts and working directories.
Allow scripts run from StdIn to use the initial working directory as the parent dir for the local library, to retrieve auto included functions.

Current implementation does not allow this. Since the behavior is to use BIV_ScriptDir, but the script's directory is empty when run from StdIn and thus the auto includes fail.

This change also lets StdIn scripts preserve A_WorkingDir as A_InitialWorkingDir since the beginning without auto-changing it to A_ScriptDir which is empty. So another script/program/caller can specify a working directory and remain valid.