Added primitive types and Any to the type hierarchy.
Added prototypes for Any, Primitive, String, Number, Integer and Float;
not yet accessible except through a value of the applicable type.
All values are based on Any (except ComObjects) and can use HasBase,
HasProp, HasMethod and GetMethod, which are now also functions.
Define/Delete methods can't be called directly on the prototypes as
they are not derived from Object.Prototype.