Removing HOT_IF_EXPR
#HotIf expr now uses HOT_IF_CALLBACK, with callback

<hotkey>(ThisHotkey) {
    return expr
}

"#HotIf var" is no longer optimized.

Reason, HOT_IF_CALLBACK can pass parameters. Removes the need (and limitation) for A_ThisHotkey.
"#HotIf var" is not optimized for brevity, questionable benefit and for consitent behaviour (scope and meaning of result of expr).

Also remove EvaluateHotCriterionExpression(), because it is not called anymore.