Revert "Changed MouseClick/Drag and L/RButton to be unaffected by button swap."
This reverts commit 16453e9fe945a3e6884d585abfdaea36d6f6e86d.

This approach was originally intended to preserve the behaviour of vk1/vk2
as "physical" L/R buttons, but using the fake VK_*_LOGICAL values is more
complicated and error-prone, such as when missing a needed translation.

Also, VK_LBUTTON/VK_RBUTTON represent the logical, not physical, buttons
in the hook, GetAsyncKeyState and maybe other places; I think it's only
the MOUSEEVENTF_ flags that are not swapped by the control panel setting.