Changed __Enum behaviour for Array, Map, Gui and RegExMatchInfo.
Mode of enumeration (which variables are assigned which values) now
depends on the first parameter of __Enum, not the passed variables.

This makes it possible for a for-loop to specify a single variable
but control which value it gets by placing a comma to either side.
Previously Array and RegExMatchInfo always returned the value,
regardless of whether the single variable was first or second.

This also changes Gui to return the control object in mode 1, instead
of the HWND (like Array and RegExMatchInfo).  Note that mode 1 is also
used by `Array(enumerable*)` and other variadic calls.