Optimized ProcessExist(PID) and ProcessWaitClose(PID) when PID exists.
If the current process has PROCESS_QUERY_LIMITED_INFORMATION|SYNCHRONIZE
access to the other process, ProcessExist may take less than 0.1% as long;
e.g. 0.001 ms instead of 2.2 ms.

In the very unlikely (but confirmed possible) event that a numeric process
name ALSO matches the PID when this shortcut is taken, the specified PID is
now preferred, whereas the PID of the other process might have been returned
if it was enumerated before the specified PID.