fix build to correctly reference PDBs from regex library
The build would generate a dozen or so errors about PDB files not
being found. This was because the build didn't reference the correct
directory for the PDB files generated by the library build. This patch
fixes that issue, and now the builds have correct symbols, as expected,
with no warnings.