fix issues in GetEnvVarReliable
GetEnvVarReliable() does some casting between size_t and DWORD, which is
problematic because size_t is eight bytes on Win32.

Further, the call to GetEnvVariable() may return non-zero but still not
copy any data to the output buffer. This case can be encountered when the
buffer is precisely the length of the required output, but can happen in
other circumstances. An extra check isn't expensive and makes this
code safer by correctly catching this error condition.