Changed tray icon handling to improve visual quality after changing DPI.
If the default icons are in use, the current icon is now reloaded when
the DPI of the primary screen is changed.

Anything that causes the tray icon to be updated may now cause the tray
icon to be recreated, such as if a script has directly deleted the icon
by calling Shell_NotifyIcon with DllCall.