Fix BIF_Hotkey's Action assignment via Strings.
Restores the currently documented behavior, specifically:

"If Action is a hotkey name, its original function is used; specifically, the original function of the hotkey variant corresponding to the current HotIf criteria. This is usually used to restore a hotkey's original function after having changed it, but can be used to assign the function of a different hotkey, provided that both hotkeys use the same HotIf criteria."

Broken by 9d9cd3b. (See also Helgef PR 7931ec7.)