Fixed an issue with hotkeys not firing due to a race condition.
If a modifier hotkey such as ~*RWin:: called Send or GetKeyState too soon,
the OS could report that RWin isn't down, so the hook's modifier state
would be "corrected" and hotkeys would wrongly fire or fail to fire.
This was likely to occur only if another keyboard hook was installed
more recently than the script's own hook, since in that case the OS would
not update key state until the other hook's thread has resumed and returned.